Direct reference to target is made instead of external variable
Fixes #1 : By moving an element that's higher in the dom and then deleting a smaller one, said element's targetName pointed to the old delete element. This is likely because of how Angular handles ngRepeat - most likely to lower the total amount of DOM manipulation needed, it only deletes the top elements, and refills the old ones with the new one's data. Since it's still technically the same noded, the instance of the draggable function running on it is still the same as the old one - even though attr is different, targetName is already set to the old value.